KOKKOS: fix the defects found by static analysis
Acts on the findings from the CodeChecker/clang-tidy run at
download.lammps.org/analysis that survived verification against this
branch.  The one serious finding, a null-pointer dereference reachable
through "fix balance" with pppm/kk, was already fixed here by the
missing reset_grid() override, so what is left is smaller.

remap_3d_create_plan: the collective branch admits a rank with nsend or
nrecv equal to zero -- both halves have to be initialized for the
Alltoallv -- and then checks the resulting pointers against nullptr.
malloc(0) may legally return a null pointer, which would abort plan
creation on a healthy rank.  Ask for one extra byte, the way the
neighbouring commringlen allocations in remap.cpp already do.  The CPU
remap has the same shape and is fixed with it.

mliap_model_python_kokkos: casting this through void* to the
non-template handle type the Cython layer uses is a cast between
unrelated types whenever the host instantiation is in play, which it is
on a GPU build.  Spell it reinterpret_cast and record at the class what
the couple layer actually relies on.

dynamical_matrix_kokkos and third_order_kokkos: DynamicalMatrix::setup()
is not virtual and the Kokkos command() delegates to the base
command(), so the Kokkos setup() was never reachable.  It also omitted
neighbor->build_one(), external_force_clear and modify->setup() in the
third_order case, and duplicated the pair compute that the virtual
update_force() override already does.  Remove both; the base setup()
plus update_force() is the path that has been running all along.

comm_kokkos: grow the bonus exchange lists on count_bonus rather than
count.  Both are safe -- count_bonus is never larger -- but the check
now says what it means.

The rest are local: protected non-virtual destructors on the KokkosBase
mixins so "delete kkbase" cannot compile (a virtual destructor would
force noexcept on every derived destructor), the missing override
keywords in fix_shardlow_kokkos.h, operator= returning a reference in
the PACE spline interpolators, an int return for RegBlockKokkos::
inside_face which only ever yields 0 or 1, a signed cast in pair uf3 to
match the line above it, unused parameter names commented out, a
duplicate include, two unused using declarations and a commented-out
call left over from the qeq/reaxff port.

Not acted on, deliberately: the global using declarations in
kokkos_type.h and the meam impl headers (a qualification sweep across
every KOKKOS translation unit, not a bug fix), and the mlo/mhi
assignments in borders_device(), which are dead only because borders()
routes every non-SINGLE mode to CommBrick::borders() -- removing them
means refactoring that branch.

Verified with a KOKKOS/OpenMP build of the affected packages: the full
ctest suite passes, including PhononCommandsKokkos which covers the two
removed setup() overrides.  The two failures in this container,
Platform.file_is_readable and TextFileReaderTest.permissions, chmod a
file to 0 and expect the read to fail, which cannot happen when the
tests run as root; they are unrelated to these changes.

+++ src/KOKKOS/pair_uf3_kokkos.cpp
@@ -784,7 +784,7 @@ template <class DeviceType> void PairUF3Kokkos<DeviceType>::compute(int eflag_in
           ((int)d_neighbors_short.extent(0) != ignum)) {
     d_neighbors_short = Kokkos::View<int **, DeviceType>("UF3::neighbors_short", ignum, max_neighs);
   }
-  if (d_numneigh_short.extent(0) != ignum)
+  if ((int)d_numneigh_short.extent(0) != ignum)
     d_numneigh_short = Kokkos::View<int *, DeviceType>("UF3::numneighs_short", ignum);
   Kokkos::parallel_for(
       Kokkos::RangePolicy<DeviceType, TagPairUF3ComputeShortNeigh>(0, ignum), *this);
